Make "TotalStars" int

+++ rockit/Assets/_Project/Scripts/Runtime/Features/Stats/Shared/Services/GameStatsService.cs
@@ -4,6 +4,6 @@
     {
         public float ProgressOfPassage01;
         public float ProgressOfPassage0100;
-        public float TotalStars;
+        public int TotalStars;
     }
 }
